Add E2E test for smart transaction
