--- test/e2e/tests/smart-transactions/mock-requests-for-swap-test.ts
@@ -1,341 +1,0 @@
-import { MockttpServer } from 'mockttp';
-import { mockEthDaiTrade } from '../swaps/shared';
-import { mockMultiNetworkBalancePolling } from '../../mock-balance-polling/mock-balance-polling';
-import { mockServerJsonRpc } from '../ppom/mocks/mock-server-json-rpc';
-
-const STX_UUID = '0d506aaa-5e38-4cab-ad09-2039cb7a0f33';
-
-const GET_FEES_REQUEST_INCLUDES = {
-  txs: [
-    {
-      from: '0x5cfe73b6021e818b776b421b1c4db2474086a7e1',
-      to: '0x881D40237659C251811CEC9c364ef91dC08D300C',
-      value: '0x1bc16d674ec80000',
-      nonce: '0x0',
-    },
-  ],
-};
-
-const GET_FEES_RESPONSE = {
-  blockNumber: 20728974,
-  id: '19d4eea3-8a49-463e-9e9c-099f9d9571ca',
-  txs: [
-    {
-      cancelFees: [],
-      return: '0x',
-      status: 1,
-      gasUsed: 190780,
-      gasLimit: 239420,
-      fees: [
-        {
-          maxFeePerGas: 4667609171,
-          maxPriorityFeePerGas: 1000000004,
-          gas: 239420,
-          balanceNeeded: 1217518987960240,
-          currentBalance: 751982303082919400,
-          error: '',
-        },
-      ],
-      feeEstimate: 627603309182220,
-      baseFeePerGas: 2289670348,
-      maxFeeEstimate: 1117518987720820,
-    },
-  ],
-};
-
-const SUBMIT_TRANSACTIONS_REQUEST_EXACTLY = {
-  rawTxs: [
-    '0x02f91a3b0180843b9aca048501163610538303a73c94881d40237659c251811cec9c364ef91dc08d300c881bc16d674ec80000b919c65f575529000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001bc16d674ec8000000000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000000136b796265725377617046656544796e616d69630000000000000000000000000000000000000000000000000000000000000000000000000000000000000018e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006b175474e89094c44da98b954eedeac495271d0f0000000000000000000000000000000000000000000000001b83413f0b3640000000000000000000000000000000000000000000000000f7a8daea356aa92bfe0000000000000000000000000000000000000000000000000000000000000120000000000000000000000000000000000000000000000000003e2c284391c000000000000000000000000000f326e4de8f66a0bdc0970b79e0924e33c79f1915000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000017a4e21fd0e90000000000000000000000000000000000000000000000000000000000000020000000000000000000000000f081470f5c6fbccf48cc4e5b82dd926409dcdd67000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000132000000000000000000000000000000000000000000000000000000000000015200000000000000000000000000000000000000000000000000000000000001260000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee0000000000000000000000006b175474e89094c44da98b954eedeac495271d0f00000000000000000000000074de5d4fcbf63e00296fd95d33236b97940166310000000000000000000000000000000000000000000000000000000066e029cc00000000000000000000000000000000000000000000000000000000000012000000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000008200000000000000000000000000000000000000000000000000000000000000ac00000000000000000000000000000000000000000000000000000000000000e80000000000000000000000000000000000000000000000000000000000000000300000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000000000000000068000000000000000000000000000000000000000000000000000000000000000400e00deaa000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000080000000000000000000000000c697051d1c6296c24ae3bcef39aca743861d9a81000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20000000000000000000000007fc66500c84a76ad7e9c93437bfc5ac33e2ddae900000000000000000000000000000000000000000000000006e0d04fc2cd90000000000000000000000000000000000000000000000000000000000000000040003c5f890000000000000000000000000000000000000000000000000000000300000000000000000000000000000000000000000000000000000000000004e00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000ead050515e10fdb3540ccd6f8236c46790508a7600000000000000000000000000000000000000000000000074b3f935bb79d45400000000000000000000000000000000000000000000000000000000000000e00000000000000000000000007fc66500c84a76ad7e9c93437bfc5ac33e2ddae9000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec7000000000000000000000000f081470f5c6fbccf48cc4e5b82dd926409dcdd670000000000000000000000000000000000000000000000c4000000000000000000000000000000000000000000000000000000000000000000000000000003a4e525b10b000000000000000000000000000000000000000000000000000000000000002000000000000000000000000028cacd5e26a719f139e2105ca1efc3d9dc892826000000000000000000000000ff8ba4d1fc3762f6154cc942ccf30049a2a0cec6000000000000000000000000f081470f5c6fbccf48cc4e5b82dd926409dcdd670000000000000000000000007fc66500c84a76ad7e9c93437bfc5ac33e2ddae9000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec700000000000000000000000000000000000000000000000074b3f935bb79d45400000000000000000000000000000000000000000000000074b3f935bb79d4540000000000000000000000000000000000000000000000000000000045aff3b30000000000000000000000000000000000000000000000000000000066e025760000000000000000000000000000000000000000000000000016649acb241b017da53b79cbf14cc2a737cd6469098549000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000026000000000000000000000000067297ee4eb097e072b4ab6f1620268061ae80464000000000000000000000000ae4fdcc420f1409c8b9b2af04db150dd986f66a5000000000000000000000000000000000000000000000000000000000000000500000000000000000000000000000000000000000000000000000000000002e0000000000000000000000000000000000000000000000000000000000000036000000000000000000000000000000000000000000000000000000000444b27250000000000000000000000000000000000000000000000000000000000000041031bc9026b766621ebb870691407a8f5b5d222977566d0bb38bbd633459fc9671e24b5c970373555d66f0a46e830ee1605152bd519fed1a9684a097364f8b41f1b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000413e4923699eff11cb0252c3f8b42793eeac8793bea92843fa4028b80ff3391bbf1df4ddef51732ceeb6f65a8c9dc2651e4b952568d350b4029d4b8b5cae5c1f991c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000408cc7a56b0000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c879c58f70905f734641735bc61e45c19dd9ad60bc0000000000000000000004e7000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000006b175474e89094c44da98b954eedeac495271d0f0000000000000000000000000000000000000000000000000000000045921fcd000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000180000000000000000000000000000000000000000000000000000000000000004063407a490000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000f081470f5c6fbccf48cc4e5b82dd926409dcdd6700000000000000000000000011b815efb8f581194ae79006d24e0d814b7697f6000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000000000000000000000000000000840f9f95029e0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000408cc7a56b0000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c879c58f70905f734641735bc61e45c19dd9ad60bc0000000000000000000004e7000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000006b175474e89094c44da98b954eedeac495271d0f000000000000000000000000000000000000000000000000000000005354532a0000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000000000002a0000000000000000000000000000000000000000000000000000000000000004063407a490000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000f081470f5c6fbccf48cc4e5b82dd926409dcdd6700000000000000000000000011b815efb8f581194ae79006d24e0d814b7697f6000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000000000000000000000000000000840f9f95029e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000040301a40330000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000a00000000000000000000000005b6a0771c752e35b2ca2aff4f22a66b1598a2bc50000000000000000000000000000000000000000000000000000000053516d7f000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec7000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48000000000000000000000000f081470f5c6fbccf48cc4e5b82dd926409dcdd6700000000000000000000000000000000000000000000000000000000000000408cc7a56b0000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c879c58f70905f734641735bc61e45c19dd9ad60bc0000000000000000000004e7000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb480000000000000000000000006b175474e89094c44da98b954eedeac495271d0f000000000000000000000000000000000000000000000000000000005351dd8d000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000180000000000000000000000000000000000000000000000000000000000000004063407a490000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000f081470f5c6fbccf48cc4e5b82dd926409dcdd6700000000000000000000000011b815efb8f581194ae79006d24e0d814b7697f6000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec700000000000000000000000000000000000000000000000004207cfca814f0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000408cc7a56b0000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c806df3b2bbb68adc8b0e302443692037ed9f91b42000000000000000000000063000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000006b175474e89094c44da98b954eedeac495271d0f0000000000000000000000000000000000000000000000000000000029a7a0aa0000000000000000000000000000000000000000000000000000000000000020000000000000000000108fd5cc11eaa000000000000000fcb6c0091c62637b42000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee0000000000000000000000006b175474e89094c44da98b954eedeac495271d0f0000000000000000000000000000000000000000000000000000000000000160000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000000000001c000000000000000000000000074de5d4fcbf63e00296fd95d33236b97940166310000000000000000000000000000000000000000000000001b83413f0b3640000000000000000000000000000000000000000000000000f7a8daea356aa92bfe000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001e00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002337b22536f75726365223a226d6574616d61736b222c22416d6f756e74496e555344223a22343635382e373832393833313733303632222c22416d6f756e744f7574555344223a22343637362e393836303733303034303236222c22526566657272616c223a22222c22466c616773223a302c22416d6f756e744f7574223a2234363631373438303431393032373532373538353934222c2254696d657374616d70223a313732353936353539362c22496e74656772697479496e666f223a7b224b65794944223a2231222c225369676e6174757265223a22546363534e7837537235376b367242794f5a74344b714472344d544637356b7651527658644230724266386e395864513869634a3830385963355155595a34675a52527645337777433237352f59586a722f34625065662b4a58514b4969556b6334356a4e73556c366e6141387141774d5a48324f4a3234657932647253386c52625551444f67784b4d6979334d413164467472575241306f6d6e664873365044624b6d6f4e494c58674b45416e497a6b6d687a675043346e396d39715043337a457459737875457042772b386356426b684e7761684f56625850635854646977334870437356365555635375522f4a495342386d6a737a494b6d664b46595a716333516c5a714e6e507a50576a3648366e73587050512b6145725338334c3544554b5868364e6a70584855764748314d7a557074584169615634737354795849582f435645685a396e76564845746b2f776b6a42673d3d227d7d0000000000000000000000000000000000000000000000000000000000000000000000000000000000000ac001a0e067d3acdb151721e7fdb3834cf2563d667aad9b4c18a5afb81390d6288ac2fe9fa0e74a9e40e017bcd926f3c5da4355e0926ae7e45b3b9e1bc474507220cb43',
-  ],
-  rawCancelTxs: [],
-};
-
-const GET_BATCH_STATUS_RESPONSE_PENDING = {
-  '0d506aaa-5e38-4cab-ad09-2039cb7a0f33': {
-    cancellationFeeWei: 0,
-    cancellationReason: 'not_cancelled',
-    deadlineRatio: 0,
-    isSettled: false,
-    minedTx: 'not_mined',
-    wouldRevertMessage: null,
-    minedHash: '',
-    timedOut: false,
-    proxied: false,
-    type: 'sentinel',
-  },
-};
-
-const GET_BATCH_STATUS_RESPONSE_SUCCESS = {
-  '0d506aaa-5e38-4cab-ad09-2039cb7a0f33': {
-    cancellationFeeWei: 0,
-    cancellationReason: 'not_cancelled',
-    deadlineRatio: 0,
-    isSettled: true,
-    minedTx: 'success',
-    wouldRevertMessage: null,
-    minedHash:
-      '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-    timedOut: true,
-    proxied: false,
-    type: 'sentinel',
-  },
-};
-
-const GET_TRANSACTION_RECEIPT_RESPONSE = {
-  id: 2901696354742565,
-  jsonrpc: '2.0',
-  result: {
-    blockHash:
-      '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-    blockNumber: '0x2',
-    contractAddress: null,
-    cumulativeGasUsed: '0xc138b1',
-    effectiveGasPrice: '0x1053fcd93',
-    from: '0x5cfe73b6021e818b776b421b1c4db2474086a7e1',
-    gasUsed: '0x2e93c',
-    logs: [
-      {
-        address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x00000000000000000000000000000000000000000000000000005af3107a4000',
-        logIndex: '0xde',
-        removed: false,
-        topics: [
-          '0xe1fffcc4923d04b559f4d29a8bfc6cda04eb5b0d3c460751c2402c5c5cc9109c',
-          '0x00000000000000000000000074de5d4fcbf63e00296fd95d33236b9794016631',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-      {
-        address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x00000000000000000000000000000000000000000000000000005a275669d200',
-        logIndex: '0xdf',
-        removed: false,
-        topics: [
-          '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-          '0x00000000000000000000000074de5d4fcbf63e00296fd95d33236b9794016631',
-          '0x00000000000000000000000051c72848c68a965f66fa7a88855f9f7784502a7f',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-      {
-        address: '0x6b175474e89094c44da98b954eedeac495271d0f',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x000000000000000000000000000000000000000000000000033dd7a160e2a300',
-        logIndex: '0xe0',
-        removed: false,
-        topics: [
-          '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-          '0x00000000000000000000000051c72848c68a965f66fa7a88855f9f7784502a7f',
-          '0x00000000000000000000000074de5d4fcbf63e00296fd95d33236b9794016631',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-      {
-        address: '0x6b175474e89094c44da98b954eedeac495271d0f',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x00000000000000000000000000000000000000000000000000006a3845cef618',
-        logIndex: '0xe1',
-        removed: false,
-        topics: [
-          '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-          '0x00000000000000000000000051c72848c68a965f66fa7a88855f9f7784502a7f',
-          '0x000000000000000000000000ad30f7eebd9bd5150a256f47da41d4403033cdf0',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-      {
-        address: '0xd82fa167727a4dc6d6f55830a2c47abbb4b3a0f8',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x0000000000000000000000006b175474e89094c44da98b954eedeac495271d0f000000000000000000000000000000000000000000000000033dd7a160e2a3000000000000000000000000000000000000000000000000000000000000000005000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc200000000000000000000000000000000000000000000000000005a275669d200',
-        logIndex: '0xe2',
-        removed: false,
-        topics: [
-          '0xb651f2787ff61b5ab14f3936f2daebdad3d84aeb74438e82870cc3b7aee71e90',
-          '0x00000000000000000000000000000000000000000000000000000191e0cc96ac',
-          '0x00000000000000000000000051c72848c68a965f66fa7a88855f9f7784502a7f',
-          '0x00000000000000000000000074de5d4fcbf63e00296fd95d33236b9794016631',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-      {
-        address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x000000000000000000000000000000000000000000000000000000cbba106e00',
-        logIndex: '0xe3',
-        removed: false,
-        topics: [
-          '0x7fcf532c15f0a6db0bd6d0e038bea71d30d808c7d98cb3bf7268a95bf5081b65',
-          '0x00000000000000000000000074de5d4fcbf63e00296fd95d33236b9794016631',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-      {
-        address: '0xf326e4de8f66a0bdc0970b79e0924e33c79f1915',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x000000000000000000000000000000000000000000000000000000cbba106e00',
-        logIndex: '0xe4',
-        removed: false,
-        topics: [
-          '0x3d0ce9bfc3ed7d6862dbb28b2dea94561fe714a1b4d019aa8af39730d1ad7c3d',
-          '0x00000000000000000000000074de5d4fcbf63e00296fd95d33236b9794016631',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-      {
-        address: '0x6b175474e89094c44da98b954eedeac495271d0f',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x000000000000000000000000000000000000000000000000033dd7a160e2a300',
-        logIndex: '0xe5',
-        removed: false,
-        topics: [
-          '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-          '0x00000000000000000000000074de5d4fcbf63e00296fd95d33236b9794016631',
-          '0x0000000000000000000000005cfe73b6021e818b776b421b1c4db2474086a7e1',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-      {
-        address: '0x881d40237659c251811cec9c364ef91dc08d300c',
-        blockHash:
-          '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-        blockNumber: '0x13c80ce',
-        data: '0x',
-        logIndex: '0xe6',
-        removed: false,
-        topics: [
-          '0xbeee1e6e7fe307ddcf84b0a16137a4430ad5e2480fc4f4a8e250ab56ccd7630d',
-          '0x015123c6e2552626efe611b6c48de60d080a6650860a38f237bc2b6f651f79d1',
-          '0x0000000000000000000000005cfe73b6021e818b776b421b1c4db2474086a7e1',
-        ],
-        transactionHash:
-          '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-        transactionIndex: '0x2f',
-      },
-    ],
-    logsBloom:
-      '0x00000000000000001000000000000000000000000000000000000001000000000000010000000000000010000000000002000000080008000000040000000000a00000000000000000020008000000000000000000540000000004008020000010000000000000000000000000000801000000000000040000000010004010000000021000000000000000000000000000020041000100004020000000000000000000000200000000000040000000000000000000000000000000000000000000000002000400000000000000000000001002000400000000000002000000000020200000000400000000800000000000000000020200400000000000001000',
-    status: '0x1',
-    to: '0x881d40237659c251811cec9c364ef91dc08d300c',
-    transactionHash:
-      '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-    transactionIndex: '0x2f',
-    type: '0x2',
-  },
-};
-
-const GET_TRANSACTION_BY_HASH_RESPONSE = {
-  id: 2901696354742565,
-  jsonrpc: '2.0',
-  result: {
-    accessList: [],
-    blockHash:
-      '0xe90b92d004a9c22c32c50c628bbd93f22e3468ec4ffc62422d68cf6370f59f1d',
-    blockNumber: '0x2',
-    chainId: '0x539',
-    from: '0x5cfe73b6021e818b776b421b1c4db2474086a7e1',
-    gas: '0x3a73c',
-    gasPrice: '0x1053fcd93',
-    hash: '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-    input:
-      '0x5f5755290000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005af3107a400000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000001c616972737761704c696768743446656544796e616d696346697865640000000000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000191e0cc96ac0000000000000000000000000000000000000000000000000000000066e44f2c00000000000000000000000051c72848c68a965f66fa7a88855f9f7784502a7f0000000000000000000000006b175474e89094c44da98b954eedeac495271d0f000000000000000000000000000000000000000000000000033dd7a160e2a300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005a275669d200000000000000000000000000000000000000000000000000000000000000001bc1acb8a206598705baeb494a479a8af9dc3a9f9b7bd1ce9818360fd6f603cf0766e7bdc77f9f72e90dcd9157e007291adc6d3947e9b6d89ff412c5b54f9a17f1000000000000000000000000000000000000000000000000000000cbba106e00000000000000000000000000f326e4de8f66a0bdc0970b79e0924e33c79f1915000000000000000000000000000000000000000000000000000000000000000000d7',
-    maxFeePerGas: '0x14bdcd619',
-    maxPriorityFeePerGas: '0x3b9aca04',
-    nonce: '0x127',
-    r: '0x5a5463bfe8e587ee1211be74580c74fa759f8292f37f970033df4b782f5e097d',
-    s: '0x50e403a70000b106e9f598b1b3f55b6ea9d2ec21d9fc67de63eb1d07df2767dd',
-    to: '0x881d40237659c251811cec9c364ef91dc08d300c',
-    transactionIndex: '0x2f',
-    type: '0x2',
-    v: '0x0',
-    value: '0x5af3107a4000',
-    yParity: '0x0',
-  },
-};
-
-export const mockSwapRequests = async (mockServer: MockttpServer) => {
-  await mockMultiNetworkBalancePolling(mockServer);
-
-  await mockServerJsonRpc(mockServer, [
-    ['eth_blockNumber'],
-    ['eth_getBlockByNumber'],
-    ['eth_chainId', { result: `0x1` }],
-  ]);
-  await mockEthDaiTrade(mockServer);
-
-  await mockServer
-    .forPost('https://transaction.api.cx.metamask.io/networks/1/getFees')
-    .withJsonBodyIncluding(GET_FEES_REQUEST_INCLUDES)
-    .thenJson(200, GET_FEES_RESPONSE);
-
-  await mockServer
-    .forPost(
-      'https://transaction.api.cx.metamask.io/networks/1/submitTransactions',
-    )
-    .once()
-    .withJsonBody(SUBMIT_TRANSACTIONS_REQUEST_EXACTLY)
-    .thenJson(200, { uuid: STX_UUID });
-
-  await mockServer
-    .forGet('https://transaction.api.cx.metamask.io/networks/1/batchStatus')
-    .withQuery({ uuids: STX_UUID })
-    .once()
-    .thenJson(200, GET_BATCH_STATUS_RESPONSE_PENDING);
-
-  await mockServer
-    .forGet('https://transaction.api.cx.metamask.io/networks/1/batchStatus')
-    .withQuery({ uuids: STX_UUID })
-    .once()
-    .thenJson(200, GET_BATCH_STATUS_RESPONSE_SUCCESS);
-
-  await mockServer
-    .forJsonRpcRequest({
-      method: 'eth_getTransactionReceipt',
-      params: [
-        '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-      ],
-    })
-    .thenJson(200, GET_TRANSACTION_RECEIPT_RESPONSE);
-
-  await mockServer
-    .forJsonRpcRequest({
-      method: 'eth_getTransactionByHash',
-      params: [
-        '0xec9d6214684d6dc191133ae4a7ec97db3e521fff9cfe5c4f48a84cb6c93a5fa5',
-      ],
-    })
-    .thenJson(200, GET_TRANSACTION_BY_HASH_RESPONSE);
-};
